backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
# psycopg2-binary==2.9.9  # Uncomment if using PostgreSQL
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic[email]==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
langchain==0.3.7
langchain-openai==0.2.6
langchain-anthropic==0.2.4
langchain-google-genai==2.0.6
langgraph==0.2.45
python-dotenv==1.0.1
